+++ src/main/java/tn/esprit/bambinou/Service/IuserService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface IuserService {
     public List<User> retrieveAllUsers();
-    public User retrieveUser(Long userId);
+    public User retrieveUser(int userId);
     public User addUser(User user);
-    public void removeUser(Long userId);
+    public void removeUser(int userId);
     public User modifyUser(User user);
 }
